pyproject.toml
[tool.poetry]
name = "logging-playground"
version = "0.1.0"
description = ""
authors = ["rojoyin <rojoyin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
